+++ src/types/settings.ts
@@ -2,7 +2,6 @@ export interface AppSettings {
   download_dir: string;
   concurrent: number;
   autoupdate_ytdlp: boolean;
-  hardware_accel: boolean;
   youtube_cookies: string;
   youtube_cookies_file: string;
   filename_template?: string | null;